apiserver/requirements.txt
google-api-core
google-cloud-run
google-cloud-pubsub
google-api-core
google-cloud-logging
fastapi
uvicorn
pytest
pytest-asyncio
pytest-mock
httpx